backend_fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
bleach==6.1.0
pytest==7.4.3
httpx==0.25.2
cryptography==41.0.7
psutil==5.9.6
pyyaml==6.0.1
